great Bot

+++ bot_obuv/handlers/mesage_hendler.py
@@ -6,7 +6,7 @@ from bot_obuv.keyboard.inline_keyboard import contact_and_address_kb
 from bot_obuv.data_base.data_base_main import update_user_visits
 
 list_price_text = ['цен', 'стоит', 'почем', 'размер']
-list_word_address = ['ехать', 'находит', 'найти', 'адрес', 'телеф', 'связ', 'звон']
+
 
 router = Router()
 
@@ -23,7 +23,7 @@ async def woman_shoes(message: Message):
     await message.delete()
     if reply:
         await reply.delete()
-        await update_user_visits(message.from_user.full_name)
+        await update_user_visits(message.from_user.full_name, message.from_user.id)
 
 
 @router.message(lambda message: 'муж' in message.text.lower())
@@ -34,7 +34,7 @@ async def men_shoes(message: Message):
     await message.delete()
     if reply:
         await reply.delete()
-        await update_user_visits(message.from_user.full_name)
+        await update_user_visits(message.from_user.full_name, message.from_user.id)
 
 
 @router.message(lambda message: 'дет' in message.text.lower())
@@ -45,7 +45,7 @@ async def kid_shoes(message: Message):
     await message.delete()
     if reply:
         await reply.delete()
-        await update_user_visits(message.from_user.full_name)
+        await update_user_visits(message.from_user.full_name, message.from_user.id)
 
 
 @router.message(lambda message: 'тап' in message.text.lower())
@@ -56,7 +56,7 @@ async def slippers(message: Message):
     await message.delete()
     if reply:
         await reply.delete()
-        await update_user_visits(message.from_user.full_name)
+        await update_user_visits(message.from_user.full_name, message.from_user.id)
 
 
 @router.message(lambda message: 'работа' in message.text.lower())
@@ -69,26 +69,7 @@ async def job_time(message: Message):
         await asyncio.sleep(15)
         await message.delete()
         await rep.delete()
-        await update_user_visits(message.from_user.full_name)
-
-
-@router.message(F.text == 'Помощь.')
-async def slippers(message: Message):
-    rep = await message.reply('Это небольшая инструкция, как пользоваться помощником в чате. '
-                              'Внизу Вы видите кнопки с разделами, нажимая на них Вы переходите в соответствующие разделы, '
-                              'позволяющие выбрать интересующий Вас род и вид обуви, посмотреть фото и цены '
-                              '(без учета скидочной карты покупателя на кожаную обувь). Все фото и цены актуальны! '
-                              'Автоматически выводится наличие товара по размерам на тот момент, когда Вы запрашиваете информацию. '
-                              'Если Вас интересует определенная модель обуви, то в поле сообщения наберите номер модели в формате "00-000", '
-                              'который находится на фото и выглядит, например, так: "99-675". Или можно просто набрать сообщение,'
-                              'допустим: "Мужские туфли", сразу выйдет меню с разделами мужской обуви:')
-
-    await message.answer('Главное меню', reply_markup=main_kb())
-    if rep:
-        await asyncio.sleep(45)
-        await rep.delete()
-        await message.delete()
-        await update_user_visits(message.from_user.full_name)
+        await update_user_visits(message.from_user.full_name, message.from_user.id)
 
 
 @router.message(F.text == 'Спасибо')
@@ -98,19 +79,7 @@ async def reply_to_thank_you(message: Message):
         await asyncio.sleep(10)
         await rep.delete()
         await message.delete()
-        await update_user_visits(message.from_user.full_name)
-
-
-@router.message(lambda message: any(word in message.text.lower() for word in list_word_address))
-async def search_type_shoes(message: Message):
-    rep = await message.reply('Если вы хотите узнать наш адрес и контакты '
-                              'нажмите нужную кнопку под сообщением', reply_markup=contact_and_address_kb)
-    await message.answer('Главное меню', reply_markup=main_kb())
-    await asyncio.sleep(20)
-    await message.delete()
-    if rep:
-        await rep.delete()
-        await update_user_visits(message.from_user.full_name)
+        await update_user_visits(message.from_user.full_name, message.from_user.id)
 
 
 @router.message(lambda message: any(word in message.text.lower() for word in list_price_text))
@@ -123,16 +92,18 @@ async def price(message: Message):
     await message.delete()
     if rep:
         await rep.delete()
-        await update_user_visits(message.from_user.full_name)
+        await update_user_visits(message.from_user.full_name, message.from_user.id)
 
 
 @router.message()
 async def no_answer(message: Message):
-    if len(message.text) != 6:
+    if len(message.text) >= 9 and len(message.text) != 6:
         rep = await message.reply('Извините я не понял вопрос')
         await message.answer('Главное меню', reply_markup=main_kb())
         if rep:
             await asyncio.sleep(10)
             await rep.delete()
             await message.delete()
-            await update_user_visits(message.from_user.full_name)
+            await update_user_visits(message.from_user.full_name, message.from_user.id)
+
+
